refactor: add useCartsState
