+++ src/pages/carts/CartsPage.tsx
@@ -1,33 +1,11 @@
-import { useGetCarts } from '@api';
 import { PageTitle } from '@components/layout';
 import Carts from './components/Carts';
 import CartAllSelector from './components/CartsAllSelector';
-import { useEffect, useState } from 'react';
-import { Cart } from '@types';
 import CartsSummary from './components/CartsSummary';
+import useCartsState from './hooks/useCartsState';
 
 export default function CartsPage() {
-  const { data } = useGetCarts();
-  const [carts, setCarts] = useState<Cart[]>([]);
-  useEffect(() => {
-    if (!data) return;
-    if (carts.length === 0) {
-      const initialCarts = (data ?? []).map((cart) => ({
-        ...cart,
-        selected: true,
-      }));
-      setCarts(initialCarts);
-      return;
-    }
-    setCarts((prev) =>
-      data.map((cart) => ({
-        ...cart,
-        selected: prev.find((prevCart) => prevCart.id === cart.id)
-          ? prev.find((prevCart) => prevCart.id === cart.id)?.selected ?? true
-          : false,
-      })),
-    );
-  }, [data]);
+  const [carts, setCarts] = useCartsState();
 
   const selectedAll = carts.every((cart) => cart.selected);
   const selectAll = () => {
